responsive-home-page

+++ src/Pages/MyProperties.jsx
@@ -1,31 +1,44 @@
 
 
-import { Link, useLoaderData } from "react-router";
+import { Link, useLoaderData, useNavigate } from "react-router";
+import Swal from "sweetalert2";
 
 
 
 export default function MyProperties() {
   const properties = useLoaderData(); 
-  
+  const navigate = useNavigate()
   const handleDelete = (id) => {
-    fetch(`http://localhost:3000/models/${id}`, {
+
+Swal.fire({
+  title: "Are you sure?",
+  text: "You won't be able to revert this!",
+  icon: "warning",
+  showCancelButton: true,
+  confirmButtonColor: "#3085d6",
+  cancelButtonColor: "#d33",
+  confirmButtonText: "Yes, delete it!"
+}).then((result) => {
+  if (result.isConfirmed) {
+        fetch(`http://localhost:3000/models/${id}`, {
   method: "DELETE",
   headers: {
     "Content-Type": "application/json",
    
   },
 })
-  .then((res) => {
-    if (!res.ok) throw new Error("Failed to delete");
-    return res.json();
-  })
-  .then((data) => console.log("Deleted:", data))
-  .catch((err) => console.error(err));
+
+  navigate('/')
+    Swal.fire({
+      title: "Deleted!",
+      text: "Your file has been deleted.",
+      icon: "success"
+    });
+  }
+});
 
 
-
-
-  };
+  }
 
   return (
     <div className="min-h-screen bg-gray-50 py-10 px-5">
